feat(contributions): add logic to handle and store a contribution validator
