+++ src/domain/entities/contribution.rs
@@ -1,6 +1,11 @@
-use crate::domain::*;
-use std::str::FromStr;
+use anyhow::anyhow;
+use starknet::core::types::FieldElement;
+use std::{fmt, str::FromStr};
 use url::Url;
+
+use super::{ContributorId, Project, ProjectId};
+
+pub type ValidatorAddress = FieldElement;
 
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub enum Status {
@@ -23,6 +28,7 @@ pub struct Contribution {
 	pub external_link: Option<Url>,
 	pub gate: u8,
 	pub metadata: Metadata,
+	pub validator: ValidatorAddress,
 }
 
 #[derive(Debug, PartialEq, Eq, Clone)]
@@ -34,19 +40,68 @@ pub struct Metadata {
 	pub r#type: Option<String>,
 }
 
-impl FromStr for Status {
-	type Err = Error;
+#[derive(Debug, PartialEq, Eq, Clone, Default)]
+pub struct Filter {
+	pub contributor_id: Option<String>,
+	pub project: Option<Project>,
+}
 
-	fn from_str(s: &str) -> Result<Self> {
+impl fmt::Display for Status {
+	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+		match self {
+			Status::Open => write!(f, "OPEN"),
+			Status::Assigned => write!(f, "ASSIGNED"),
+			Status::Completed => write!(f, "COMPLETED"),
+			Status::Abandoned => write!(f, "ABANDONED"),
+		}
+	}
+}
+
+impl FromStr for Status {
+	type Err = anyhow::Error;
+
+	fn from_str(s: &str) -> Result<Self, Self::Err> {
 		match s {
 			"OPEN" => Ok(Status::Open),
 			"ASSIGNED" => Ok(Status::Assigned),
 			"COMPLETED" => Ok(Status::Completed),
 			"ABANDONED" => Ok(Status::Abandoned),
-			_ => Err(Error::ParseStatusError(format!(
-				"Unable to parse {} into a contribution status",
-				s,
-			))),
+			_ => Err(anyhow!("Unable to parse {} into a PR status", s)),
+		}
+	}
+}
+
+#[cfg(test)]
+mod test {
+	use super::*;
+
+	#[test]
+	fn contribution_status_from_string() {
+		assert_eq!(Status::Open, "OPEN".parse().unwrap());
+		assert_eq!(Status::Assigned, "ASSIGNED".parse().unwrap());
+		assert_eq!(Status::Completed, "COMPLETED".parse().unwrap());
+		assert_eq!(Status::Abandoned, "ABANDONED".parse().unwrap());
+
+		assert!("NON_EXISTENT".parse::<Status>().is_err());
+	}
+
+	#[test]
+	fn contribution_status_to_string() {
+		assert_eq!("OPEN", Status::Open.to_string());
+		assert_eq!("ASSIGNED", Status::Assigned.to_string());
+		assert_eq!("COMPLETED", Status::Completed.to_string());
+		assert_eq!("ABANDONED", Status::Abandoned.to_string());
+	}
+
+	#[test]
+	fn contribution_status_serde() {
+		for status in [
+			Status::Open,
+			Status::Assigned,
+			Status::Completed,
+			Status::Abandoned,
+		] {
+			assert_eq!(status, status.to_string().parse().unwrap());
 		}
 	}
 }
