+++ src/domain/entities/mod.rs
@@ -1,5 +1,8 @@
 mod contribution;
-pub use contribution::{Contribution, Id as ContributionId, Status as ContributionStatus};
+pub use contribution::{
+	Contribution, Id as ContributionId, Metadata as ContributionMetadata,
+	Status as ContributionStatus,
+};
 
 mod project;
 pub use project::{Id as ProjectId, Project};
