+++ src/domain/types/mod.rs
@@ -1,3 +1,2 @@
 pub mod contracts_update_status;
-pub mod contribution;
 pub mod project;
