+++ src/routes/contributions.rs
@@ -1,4 +1,7 @@
-use std::sync::{Arc, RwLock};
+use std::{
+	str::FromStr,
+	sync::{Arc, RwLock},
+};
 
 use crypto_bigint::U256;
 use deathnote_contributions_feeder::{
@@ -14,6 +17,7 @@ use rocket::{
 	State,
 };
 use rocket_okapi::{openapi, JsonSchema};
+use starknet::core::types::FieldElement;
 
 use crate::action_queue::ActionQueue;
 
@@ -21,23 +25,31 @@ use super::ApiKey;
 
 #[derive(Deserialize, JsonSchema)]
 #[serde(crate = "rocket::serde")]
-pub struct CreateContributionFromGitHubBody {
-	issue_number: u128,
+pub struct CreateContributionBody {
+	github_issue_number: u128,
 	project_id: u128,
 	gate: u8,
+	validator: String,
 }
 
 #[openapi(tag = "Contributions")]
-#[post("/contribution/github", format = "application/json", data = "<body>")]
+#[post("/contributions/github", format = "application/json", data = "<body>")]
 pub async fn create_contribution(
 	_api_key: ApiKey,
-	body: Json<CreateContributionFromGitHubBody>,
+	body: Json<CreateContributionBody>,
 	github_api: &State<github::API>,
 	queue: &State<Arc<RwLock<ActionQueue>>>,
 ) -> Result<Status, Json<HttpApiProblem>> {
 	let body = body.into_inner();
+	let validator = FieldElement::from_str(&body.validator).map_err(|e| {
+		Json(
+			HttpApiProblem::new(StatusCode::BAD_REQUEST)
+				.title("Invalid validator address")
+				.detail(e.to_string()),
+		)
+	})?;
 
-	let github_issue = github_api.issue(body.project_id, body.issue_number).await;
+	let github_issue = github_api.issue(body.project_id, body.github_issue_number).await;
 	let github_issue = match github_issue {
 		Ok(github_issue) => github_issue,
 		Err(error) =>
@@ -51,7 +63,7 @@ pub async fn create_contribution(
 	let metadata = github::extract_metadata(github_issue.clone());
 
 	let contribution = domain::Contribution {
-		id: (body.project_id * 1_000_000 + body.issue_number).to_string(),
+		id: (body.project_id * 1_000_000 + body.github_issue_number).to_string(),
 		project_id: body.project_id.to_string(),
 		contributor_id: None,
 		title: Some(github_issue.title),
@@ -60,6 +72,7 @@ pub async fn create_contribution(
 		external_link: Some(github_issue.html_url),
 		gate: body.gate,
 		metadata,
+		validator,
 	};
 
 	match queue.write() {
@@ -85,7 +98,7 @@ pub struct AssignContributorBody {
 
 #[openapi(tag = "Contributions")]
 #[post(
-	"/contribution/<contribution_id>/contributor",
+	"/contributions/<contribution_id>/contributor",
 	format = "application/json",
 	data = "<body>"
 )]
@@ -117,7 +130,7 @@ pub async fn assign_contributor(
 }
 
 #[openapi(tag = "Contributions")]
-#[post("/contribution/<contribution_id>/validate")]
+#[post("/contributions/<contribution_id>/validate")]
 pub async fn validate_contribution(
 	_api_key: ApiKey,
 	contribution_id: ContributionId,
@@ -137,7 +150,7 @@ pub async fn validate_contribution(
 }
 
 #[openapi(tag = "Contributions")]
-#[delete("/contribution/<contribution_id>/contributor")]
+#[delete("/contributions/<contribution_id>/contributor")]
 pub async fn unassign_contributor(
 	_api_key: ApiKey,
 	contribution_id: ContributionId,
